* Settings screen allows for PTT selection.  * Added a server node in the TreeView
git-svn-id: http://subversion.assembla.com/svn/gablarski/trunk@263 080cfd25-d5d9-4c59-82ea-70d02a200d8b

+++ Clients/Windows/PushToTalk.cs
@@ -43,7 +43,7 @@ namespace Gablarski.Clients.Windows
 		public override string ToString ()
 		{
 			//return (keyboard) ? "k" + KeyboardKeys : "m" + MouseButtons;
-			return "k" + KeyboardKeys;
+			return "k" + ((int)KeyboardKeys);
 		}
 
 		public static PushToTalk Parse (string value)
